moved hpa metric test
